End of Milestone 1

+++ src/ViewFlights.js
@@ -27,7 +27,7 @@ class ViewFlights extends Component {
 
     }
    
-
+  
     componentDidMount() {
         axios.get('/allFlights')
             .then(res => {
@@ -41,6 +41,8 @@ class ViewFlights extends Component {
         this.props.navigate('/updateFlight/'+deletedFlightNumber)
     }
     deleteFlight(deletedFlightNumber) {
+    
+
         axios({
             method: "delete",
             url: "/flight/" + deletedFlightNumber,
@@ -116,7 +118,12 @@ class ViewFlights extends Component {
       <button
                       className="bg-blueGray-800 text-white active:bg-blueGray-600 text-sm font-bold uppercase px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 w-full ease-linear transition-all duration-150"
                       type="button"  onClick={(e) =>{ e.preventDefault();
-                        this.deleteFlight(fl.FlightNumber);
+                        if (window.confirm("Press a button!")) {
+                         this.deleteFlight(fl.FlightNumber);
+
+                      } else {
+                        
+                        }
                         
                             }
                     }
